requirements.txt
paho-mqtt
CherryPy
autopep8
pylint